requirements.txt
aioprometheus==21.9.0
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
Automat==24.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
decorator==5.1.1
fastapi==0.95.1
flake8==7.1.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.1
pluggy==1.5.0
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.20.0
pycodestyle==2.12.1
pydantic==1.10.9
pydantic_core==2.20.1
pyflakes==3.2.0
PyMySQL==1.0.3
pytest==8.3.2
pytest-asyncio==0.24.0
pytest-tornasync==0.6.0.post2
python-dotenv==1.0.0
quantile-python==1.1
requests==2.32.3
setuptools==73.0.1
sniffio==1.3.1
starlette==0.26.1
tornado==6.4.1
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.23.1
zope.interface==7.0.1
